Design improvements, and loading status on answers
Added some colour and design thought to the UI.

When calculating answers, we first show a loading UI and then kick off
the calculation, because it blocks the UI thread. This means we're
able to wanr the user that it is doing something.

Better still, would be to use a background worker to calculate the
answers on a separate thread.